requirements.txt
boto3==1.9.142
botocore==1.12.142
Django==2.2.1
docutils==0.14
jmespath==0.9.4
Pillow==6.0.0
python-dateutil==2.8.0
pytz==2019.1
s3transfer==0.2.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.3
